Update for new conduit, resourceT
This is for #31

A few things here:
1. Several operators and type aliases in conduit that have been
provided for backwards compatibility are now actually deprecated. I've
updated our use of type aliases and operators to the recommended path.
2. Conduit has dropped the mmorph dependency which means the hoisting
we were doing wasn't working. They aliased ExceptionT to CatchT, a
monad transformer built for providing MonadThrow for pure code which
we need because the CSV machinery requires a MonadThrow
instance. CatchT was kind of limited so I handle it just in the CSV
parsing part, convert it to an either and then change the transformer
stack to ExceptT so it aborts the pipeline if it throws.

However, I think a lot of this is all hypothetical. Looking at the
code today, I cannot see a situation where the parsing can even throw
an exception. csv-conduit takes your parser, runs it in a non-throwing
manner and then eats rows. See #24 for some efforts of fixing that. So
I couldn't write a test that exercised this case but the types check
and non-exceptional decoding appears to work correctly.